requirements.txt
streamlit
opencv-python-headless
numpy
scikit-learn
rembg
Pillow
matplotlib